+++ CHANGELOG.md
@@ -12,6 +12,18 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - New config values `password_reset_token_lifetime`
   and `email_verification_token_lifetime`: https://github.com/supertokens/supertokens-core/issues/297
 
+## [3.5.2] - 2021-09-01
+
+### Fixes
+
+- Issue with verifying refresh token throwing an unauthorised exception due to a db connection error.
+- Sends far ahead jwt signing key expiry time in case updating them is
+  disabled: https://github.com/supertokens/supertokens-core/issues/304
+
+### Changes
+
+- Changes JWT signing key update interval to not be limited to 720 hours
+
 ## [3.5.1] - 2021-08-25
 
 ### Added